inserted samples successfully to build template
