+++ app.py
@@ -1,10 +1,14 @@
 import streamlit as st
-import pandas as pd
-import requests
 import pyrebase
 from firebase_admin import auth
 import json
 from streamlit.components.v1 import html
+from authentication_functions import log_in, log_out, sign_up, reset_user_password, status
+from streamlit_option_menu import option_menu
+from sales_dashboard import sales_report
+from sankey_dashboard import sankey_report
+from scatterplot import scatterplot_report
+from bar_chart import bar_report
 
 with open('firebase_app_config.json') as f:
     config = json.load(f)
@@ -14,147 +18,95 @@ auth = firebase. auth()
 
 def main():
 
-    def log_in():
-        email = st.session_state.login_email
-        password = st.session_state.login_password
-        try:
-            user = auth.sign_in_with_email_and_password(email, password)
-            idtoken = user['idToken']
-            userinfo = auth.get_account_info(idtoken)
-            useremailverified = userinfo['users'][0]['emailVerified']
-            if useremailverified == True:
-                st.session_state['authenticated'] = True
-                st.session_state['email'] = userinfo['users'][0]['email']
-                st.session_state['user'] = user
-                st.success("{} is signed in ".format(email))
-            else:
-                #send email to user to verify email
-                auth.send_email_verification(user['idToken'])
-                st.error("Hello {}, please verify your email first, we have sent you an email".format(email))
+    st.set_page_config(page_title="Sales Dashboard", page_icon=":bar_chart:", layout="wide") #layout can be centered
 
-        except requests.HTTPError as e:
-            error_json = e.args[1]
-            error = json.loads(error_json)['error']['message']
-            st.error(error)
+    # st.title("streamlit forms")
 
-        except Exception as e:
-            st.error(e)
+    with st.sidebar:
+        selected = option_menu(
+            menu_title="Main Menu",  # required
+            options=["Home", 
+                "Chart1", 
+                "Chart2", 
+                "Chart3",
+                "Chart4",
+                "Chart5"
+                ],  # required
+            icons=["house", "book", "envelope"],  # optional
+            menu_icon="cast",  # optional
+            default_index=0,  # optional - what option is selected by default first
+        )
 
+    if selected == "Home":
+        st.title(f"You have selected {selected}")
 
+        "st.session_state object:", st.session_state
 
-    def log_out():
-        del st.session_state["user"]
-        st.session_state['authenticated'] = False
-
-    def sign_up():
-        email = st.session_state.signup_email
-        password = st.session_state.signup_password
-
-        try:
-            user = auth.create_user_with_email_and_password(email, password)
-            auth.send_email_verification(user['idToken'])
-            st.success("Hello {}, we have sent an email to you, please check and confirm".format(email))
-
-        except requests.HTTPError as e:
-            error_json = e.args[1]
-            error = json.loads(error_json)['error']['message']
-            st.error(error)
-
-        except Exception as e:
-            st.error(e)
-
-    def reset_user_password():
-        try:
-            email = st.session_state.reset_email_password
-            auth.send_password_reset_email(email)
-            st.success("Hello {}, we have sent an email to you, please click on the validation link".format(email))
-
-        except requests.HTTPError as e:
-            error_json = e.args[1]
-            error = json.loads(error_json)['error']['message']
-            st.error(error)
-
-        except Exception as e:
-            st.error(e)
-
-
-    def status():
-
-        try:
+        #checking if user token is in session, if not user is not authenticated
+        if 'user' in st.session_state.keys():
             user = st.session_state['user']
             userinfo = auth.get_account_info(user['idToken'])
-            if userinfo['users'][0]['emailVerified'] == True:
-                st.session_state['authenticated'] = True    
-                auth.refresh(user['refreshToken'])
-                useremail = userinfo['users'][0]['email']
-                st.write("{} is logged in".format(useremail))
-        except:
+            #checking if user has email verified, if not user is not authenticated
+            if userinfo['users'][0]['emailVerified'] == False:
+                st.session_state['authenticated'] = False
+            else:    
+                st.session_state['authenticated'] = True
+        else:
             st.session_state['authenticated'] = False
-            st.write("User is not logged in")
 
-
-    st.title("streamlit forms")
-    menu = ["Home", "About"]
-    choice = st.sidebar.selectbox("Menu", menu )
-
-    "st.session_state object:", st.session_state
-
-
-    #checking if user token is in session, if not user is not authenticated
-    if 'user' in st.session_state.keys():
-        user = st.session_state['user']
-        userinfo = auth.get_account_info(user['idToken'])
-        #checking if user has email verified, if not user is not authenticated
-        if userinfo['users'][0]['emailVerified'] == False:
-            st.session_state['authenticated'] = False
-        else:    
-            st.session_state['authenticated'] = True
-    else:
-        st.session_state['authenticated'] = False
-
-
-
-
-    if choice == "Home":
 
         if st.session_state['authenticated'] == False:
 
-            a = st.radio("", ['Login', 'Signup', 'Reset Email'], 0, horizontal=True)
-            if a == 'Login':
+            buffer, col2, buffer = st.columns([2,4,2])
 
+            with col2:
 
-                st.subheader("Log In")
-                with st.form(key="SignInForm", clear_on_submit=True):
-                    email = st.text_input("Enter your Email", key="login_email")
-                    password = st.text_input("Enter a password", type="password", key="login_password")
-                    submit_button = st.form_submit_button(label="Log in", on_click=log_in)
+                a = st.radio("", ['Login', 'Signup', 'Reset Email'], 0, horizontal=True)
+                if a == 'Login':
+                    st.subheader("Log In")
+                    with st.form(key="SignInForm", clear_on_submit=True):
+                        email = st.text_input("Enter your Email", key="login_email")
+                        password = st.text_input("Enter a password", type="password", key="login_password")
+                        submit_button = st.form_submit_button(label="Log in", on_click=log_in)
 
-            elif a == 'Signup':
-                st.subheader("Sign Up")
-                with st.form(key="SignUpForm", clear_on_submit=True):
-                    email = st.text_input("Enter your Email", key="signup_email")
-                    password = st.text_input("Enter a password", type="password", key="signup_password")
-                    submit_button = st.form_submit_button(label="Sign up", on_click=sign_up)
-
-            else:
-
-                st.subheader("Reset Email")
-                with st.form(key="ResetEmail", clear_on_submit=True):
-                    email = st.text_input("Enter your Email", key="reset_email_password")
-                    submit_button = st.form_submit_button(label="Reset", on_click=reset_user_password)
-
-
+                elif a == 'Signup':
+                    st.subheader("Sign Up")
+                    with st.form(key="SignUpForm", clear_on_submit=True):
+                        email = st.text_input("Enter your Email", key="signup_email")
+                        password = st.text_input("Enter a password", type="password", key="signup_password")
+                        submit_button = st.form_submit_button(label="Sign up", on_click=sign_up)
+                else:
+                    st.subheader("Reset Email")
+                    with st.form(key="ResetEmail", clear_on_submit=True):
+                        email = st.text_input("Enter your Email", key="reset_email_password")
+                        submit_button = st.form_submit_button(label="Reset", on_click=reset_user_password)
 
         else:
             submit_button = st.button(label="Logout", on_click=log_out)
 
             status()
 
-    else:
-        st.subheader("About")
 
+
+    if selected == "Chart1":
+        st.title(f"You have selected {selected}")
+        sales_report()
         status()
 
+    if selected == "Chart2":
+        st.title(f"You have selected {selected}")
+        sankey_report()
+        status()
+
+    if selected == "Chart3":
+        st.title(f"You have selected {selected}")
+        scatterplot_report()
+        status()
+
+    if selected == "Chart4":
+        st.title(f"You have selected {selected}")
+        bar_report()
+        status()
 
 
 if __name__ == '__main__':
