+++ bar_chart.py
@@ -6,78 +6,78 @@ import streamlit as st
 import plotly.express as px
 from PIL import Image
 
-
-
 db = firestore.Client.from_service_account_json("serviceAccountKey.json")
 
-st.set_page_config(page_title='Survey Results')
-st.header('Survey Results 2021')
-st.subheader('Was the tutorial helpful?')
 
-# ### --- LOAD DATAFRAME
-surveylist = []
-docs = db.collection('Survey').get()
-for doc in docs:
-    surveylist.append(doc.to_dict())
+def bar_report():
 
-#turning the data into a dataframe
-df = pd.DataFrame(surveylist)
+    st.header('Survey Results 2021')
+    st.subheader('Was the tutorial helpful?')
 
-#aggregate the data by department
-df_participants = df.groupby(['Department'])['Age'].count()
+    # ### --- LOAD DATAFRAME
+    surveylist = []
+    docs = db.collection('Survey').get()
+    for doc in docs:
+        surveylist.append(doc.to_dict())
 
-#convert to datafram from series
-df_participants = pd.DataFrame({'Department':df_participants.index, 'Count':df_participants.values})
+    #turning the data into a dataframe
+    df = pd.DataFrame(surveylist)
 
-# --- STREAMLIT SELECTION
-department = df['Department'].unique().tolist()
-ages = df['Age'].unique().tolist()
+    #aggregate the data by department
+    df_participants = df.groupby(['Department'])['Age'].count()
 
-age_selection = st.slider('Age:',
-                        min_value= min(ages),
-                        max_value= max(ages),
-                        value=(min(ages),max(ages)))
+    #convert to datafram from series
+    df_participants = pd.DataFrame({'Department':df_participants.index, 'Count':df_participants.values})
+
+    # --- STREAMLIT SELECTION
+    department = df['Department'].unique().tolist()
+    ages = df['Age'].unique().tolist()
+
+    age_selection = st.slider('Age:',
+                            min_value= min(ages),
+                            max_value= max(ages),
+                            value=(min(ages),max(ages)))
 
 
-department_selection = st.multiselect('Department:',
-                                    department,
-                                    default=department)
+    department_selection = st.multiselect('Department:',
+                                        department,
+                                        default=department)
 
-# --- FILTER DATAFRAME BASED ON SELECTION
-mask = (df['Age'].between(*age_selection)) & (df['Department'].isin(department_selection))
-number_of_result = df[mask].shape[0]
-st.markdown(f'*Available Results: {number_of_result}*')
+    # --- FILTER DATAFRAME BASED ON SELECTION
+    mask = (df['Age'].between(*age_selection)) & (df['Department'].isin(department_selection))
+    number_of_result = df[mask].shape[0]
+    st.markdown(f'*Available Results: {number_of_result}*')
 
-# --- GROUP DATAFRAME AFTER SELECTION
-df_grouped = df[mask].groupby(by=['Rating']).count()[['Age']]
-df_grouped = df_grouped.rename(columns={'Age': 'Votes'})
-df_grouped = df_grouped.reset_index()
+    # --- GROUP DATAFRAME AFTER SELECTION
+    df_grouped = df[mask].groupby(by=['Rating']).count()[['Age']]
+    df_grouped = df_grouped.rename(columns={'Age': 'Votes'})
+    df_grouped = df_grouped.reset_index()
 
-# --- PLOT BAR CHART
-bar_chart = px.bar(df_grouped,
-                   x='Rating',
-                   y='Votes',
-                   text='Votes',
-                   color_discrete_sequence = ['#F63366']*len(df_grouped),
-                   template= 'plotly_white')
-st.plotly_chart(bar_chart)
+    # --- PLOT BAR CHART
+    bar_chart = px.bar(df_grouped,
+                    x='Rating',
+                    y='Votes',
+                    text='Votes',
+                    color_discrete_sequence = ['#F63366']*len(df_grouped),
+                    template= 'plotly_white')
+    st.plotly_chart(bar_chart)
 
-# --- DISPLAY IMAGE & DATAFRAME
-col1, col2 = st.columns(2)
-image = Image.open('images/survey.jpg')
+    # --- DISPLAY IMAGE & DATAFRAME
+    col1, col2 = st.columns(2)
+    image = Image.open('images/survey.jpg')
 
-col1.image(image,
-        caption='Designed by slidesgo / Freepik',
-        use_column_width=True)
-col2.dataframe(df[mask])
+    col1.image(image,
+            caption='Designed by slidesgo / Freepik',
+            use_column_width=True)
+    col2.dataframe(df[mask])
 
-# --- PLOT PIE CHART
-pie_chart = px.pie(df_participants,
-                title='Total No. of Participants',
-                values='Count',
-                names='Department')
+    # --- PLOT PIE CHART
+    pie_chart = px.pie(df_participants,
+                    title='Total No. of Participants',
+                    values='Count',
+                    names='Department')
 
-st.plotly_chart(pie_chart)
+    st.plotly_chart(pie_chart)
