--- selectbox_scatterplot.py
@@ -1,65 +1,0 @@
-#Import the required Libraries
-
-import streamlit as st
-import pandas as pd
-import matplotlib.pyplot as plt
-import plotly.express as px
-
-
-st.set_page_config(layout="wide")
-df = pd.read_csv('data/data.csv')
-
-# Functions for each of the pages
-def home(df):
-    pass
-
-def data_summary():
-    st.header('Statistics of Dataframe')
-    st.write(df.describe())
-
-def data_header():
-    st.header('Header of Dataframe')
-    st.write(df.head())
-
-def displayplot():
-    st.header('Plot of Data')
-    
-    fig, ax = plt.subplots(1,1)
-    ax.scatter(x=df['Depth'], y=df['Magnitude'])
-    ax.set_xlabel('Depth')
-    ax.set_ylabel('Magnitude')
-    st.pyplot(fig)
-
-def interactive_plot():
-    col1, col2 = st.columns(2)
-
-    x_axis_val = col1.selectbox('Select the X-axis', options=df.columns)
-    y_axis_val = col2.selectbox('Select the Y-axis', options=df.columns)
-
-    plot = px.scatter(df, x=x_axis_val, y=y_axis_val)
-    st.plotly_chart(plot, use_container_width=True)
-
-# Add a title and intro text
-st.title('Earthquake Data Explorer')
-st.text('This is a web app to allow exploration of Earthquake Data')
-
-# Sidebar setup
-st.sidebar.title('Sidebar')
-#Sidebar navigation
-st.sidebar.title('Navigation')
-options = st.sidebar.radio('Select what you want to display:', ['Home', 'Data Summary', 'Data Header', 'Scatter Plot', 'Fancy Plots'])
-
-
-# Navigation options
-if options == 'Home':
-    home(df)
-elif options == 'Data Summary':
-    data_summary()
-elif options == 'Data Header':
-    data_header()
-elif options == 'Scatter Plot':
-    displayplot()
-elif options == 'Fancy Plots':
-    interactive_plot()
-
-# tutorial from https://youtu.be/3f-j-PZ5N8A
